Splittet ut transformeringsfunksjoner til egen fil, transform.js
